@@ -3,34 +3,27 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@emotion/react": "^11.4.1",
-    "@emotion/styled": "^11.3.0",
-    "@material-ui/core": "^4.12.2",
-    "@material-ui/data-grid": "^4.0.0-alpha.34",
-    "@material-ui/icons": "^4.11.2",
-    "@material-ui/lab": "^4.0.0-alpha.60",
-    "@material-ui/styles": "^4.11.2",
-    "@mui/icons-material": "^5.0.0",
-    "@mui/material": "^5.0.0",
-    "@testing-library/jest-dom": "^5.14.1",
-    "@testing-library/react": "^12.0.0",
-    "@testing-library/user-event": "^13.1.9",
-    "bootstrap": "^5.0.2",
-    "cors": "^2.8.5",
-    "css-what": "^5.0.1",
-    "firebase": "^9.0.2",
-    "moment": "^2.29.1",
-    "node-sass": "^6.0.0",
-    "react": "^17.0.2",
-    "react-bootstrap": "^1.6.1",
-    "react-dom": "^17.0.2",
-    "react-highlight-words": "^0.17.0",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "4.0.3",
-    "react-scroll-up": "^1.3.7",
-    "typeface-merriweather": "^1.1.13",
-    "typeface-roboto": "^1.1.13",
-    "web-vitals": "^2.1.0"
+    "@emotion/react": "^11.13.3",
+    "@emotion/styled": "^11.13.0",
+    "@mui/icons-material": "^6.1.2",
+    "@mui/material": "^6.1.2",
+    "@mui/styles": "^6.1.1",
+    "@mui/utils": "^6.1.2",
+    "@mui/x-data-grid": "^7.18.0",
+    "aos": "^2.3.4",
+    "cors": "2.8.5",
+    "firebase": "^10.14.0",
+    "moment": "^2.30.1",
+    "react": "^18.3.1",
+    "react-circular-progressbar": "^2.1.0",
+    "react-dom": "^18.3.1",
+    "react-highlight-words": "0.20.0",
+    "react-router-dom": "6.26.2",
+    "react-scripts": "^5.0.1",
+    "react-scroll-up": "1.4.0",
+    "typeface-merriweather": "1.1.13",
+    "typeface-roboto": "1.1.13",
+    "web-vitals": "^4.2.3"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -59,13 +52,20 @@
     ]
   },
   "devDependencies": {
-    "body-parser": "^1.19.0",
-    "express": "^4.17.1",
-    "express-pino-logger": "^6.0.0",
-    "node-env-run": "^4.0.2",
-    "nodemon": "^2.0.12",
-    "npm-run-all": "^4.1.5",
-    "pino-colada": "^2.1.0"
+    "@babel/plugin-proposal-private-property-in-object": "^7.21.11",
+    "body-parser": "^1.20.3",
+    "css-loader": "^7.1.2",
+    "eslint-plugin-import": "^2.30.0",
+    "express": "^4.21.0",
+    "express-pino-logger": "^7.0.0",
+    "node-env-run": "4.0.2",
+    "nodemon": "^3.1.7",
+    "npm-run-all": "4.1.5",
+    "pino-colada": "^2.2.2",
+    "postcss-loader": "^8.1.1",
+    "resolve-url-loader": "^5.0.0",
+    "sass": "^1.79.4",
+    "sass-loader": "^16.0.2"
   },
-  "buildDate": 1671625285234
+  "buildDate": 1727898477322
 }
