@@ -1,4 +1,3 @@
-import React from "react";
 import Skeleton from "@mui/material/Skeleton";
 import "./loader.scss";
 
@@ -13,17 +12,17 @@ export default function LoadingSkeleton() {
               width="24px"
               height="24px"
               variant="circular"
-              style={{ marginRight: "5px" }}
+              style={{ marginRight: "5px", backgroundColor: "var(--disabled-color)" }}
             />
-            <Skeleton animation="wave" width="70%" height="30px" />
+            <Skeleton animation="wave" width="70%" height="30px" sx={{ backgroundColor: "var(--disabled-color)"}} />
           </span>
           <Skeleton
             animation="wave"
             width="10%"
             height="8px"
-            style={{ marginTop: "10px" }}
+            style={{ marginTop: "10px",  backgroundColor: "var(--disabled-color)" }}
           />
-          <Skeleton animation="wave" width="100%" height="100%" />
+          <Skeleton animation="wave" width="100%" height="100%" sx={{backgroundColor: "var(--disabled-color)"}}/>
         </li>
         <li>
           <span style={{ display: "inline-flex", alignItems: "center" }}>
@@ -32,17 +31,17 @@ export default function LoadingSkeleton() {
               width="24px"
               height="24px"
               variant="circular"
-              style={{ marginRight: "5px" }}
+              style={{ marginRight: "5px",  backgroundColor: "var(--disabled-color)" }}
             />
-            <Skeleton animation="wave" width="70%" height="30px" />
+            <Skeleton animation="wave" width="70%" height="30px" sx={{ backgroundColor: "var(--disabled-color)"}} />
           </span>
           <Skeleton
             animation="wave"
             width="10%"
             height="8px"
-            style={{ marginTop: "10px" }}
+            style={{ marginTop: "10px",  backgroundColor: "var(--disabled-color)" }}
           />
-          <Skeleton animation="wave" width="100%" height="100%" />
+          <Skeleton animation="wave" width="100%" height="100%" sx={{ backgroundColor: "var(--disabled-color)"}} />
         </li>
         <li>
           <span style={{ display: "inline-flex", alignItems: "center" }}>
@@ -51,17 +50,17 @@ export default function LoadingSkeleton() {
               width="24px"
               height="24px"
               variant="circular"
-              style={{ marginRight: "5px" }}
+              style={{ marginRight: "5px",  backgroundColor: "var(--disabled-color)" }}
             />
-            <Skeleton animation="wave" width="70%" height="30px" />
+            <Skeleton animation="wave" width="70%" height="30px" sx={{ backgroundColor: "var(--disabled-color)"}} />
           </span>
           <Skeleton
             animation="wave"
             width="10%"
             height="8px"
-            style={{ marginTop: "10px" }}
+            style={{ marginTop: "10px",  backgroundColor: "var(--disabled-color)" }}
           />
-          <Skeleton animation="wave" width="100%" height="100%" />
+          <Skeleton animation="wave" width="100%" height="100%"  sx={{ backgroundColor: "var(--disabled-color)"}} />
         </li>
       </ul>
     </div>
