@@ -1,152 +1,207 @@
+// src/Pages/Read/Read.js
+
 import "./read.scss";
-import React, { useEffect } from "react";
-import { useLocation } from "react-router-dom";
+import { useEffect, useRef, useState } from "react";
+import { useNavigate, useParams, useSearchParams } from "react-router-dom";
 import ScrollUpButton from "../../Components/scroll-up-button";
 import ScrollToTop from "react-scroll-up";
 import Highlighter from "react-highlight-words";
 import { logEvent } from "firebase/analytics";
 import { analytics } from "../../index";
+import ReadSkeleton from "../../Components/skeleton/read-skeleton";
 
 export default function Read() {
-  const [messageToRead, setMessageToRead] = React.useState([]);
-  const location = useLocation();
+  // Extract URL parameters
+  const { date, ref } = useParams(); // :date and :ref from the URL
+  const [searchParams] = useSearchParams(); // Query parameters: q and type
+
+  // Extract query parameters
+  const searchTerm = searchParams.get("q") || "";
+  const searchType = searchParams.get("type") || "semantic"; // Default to 'semantic' if not provided
+
+  // State management
+  const [messageToRead, setMessageToRead] = useState([]);
+  const [isLoading, setIsLoading] = useState(false); // Loading state
+  const [error, setError] = useState(null); // Error state
+
+  // Ref to track if the component is mounted
+  const isMountedRef = useRef(false);
+  const navigate = useNavigate();
 
   useEffect(() => {
-    let isMounted = true;
-    if (location?.state?.date) {
-      const url =
-        "https://bsaj8zf1se.execute-api.us-east-2.amazonaws.com/prod/ReadingSermonFromMySQL?sermon=" +
-        location?.state?.date;
-      fetch(url)
-        .then((response) => response.json())
-        .then(
-          (results) => {
-            if (isMounted) {
-              setMessageToRead(results);
-              executeScroll(location?.state?.ref);
-              logEvent(analytics, "read_sermon", {
-                sermon_date: location.state?.date,
-                sermon_title: results[0]?.title,
-              });
-            }
-          },
-          (error) => {
-            console.log("error: " + error);
+    isMountedRef.current = true;
+
+    console.log(`Navigating to Read Component with date: ${date}, ref: ${ref}`);
+
+    if (!date) {
+      // If date parameter is missing, redirect to Home or show an error
+      navigate("/", { replace: true });
+      return;
+    }
+
+    const fetchSermon = async () => {
+      if (date) {
+        setIsLoading(true);
+        setError(null);
+
+        // Adjust the API URL based on whether 'ref' is present
+        const url = ref
+          ? `https://bsaj8zf1se.execute-api.us-east-2.amazonaws.com/prod/ReadingSermonFromMySQL?sermon=${date}&ref=${ref}`
+          : `https://bsaj8zf1se.execute-api.us-east-2.amazonaws.com/prod/ReadingSermonFromMySQL?sermon=${date}`;
+
+        try {
+          const response = await fetch(url);
+          if (!response.ok) {
+            throw new Error(`HTTP error! status: ${response.status}`);
           }
-        );
-    }
-    return () => {
-      isMounted = false;
+          const results = await response.json();
+
+          if (isMountedRef.current) {
+            setMessageToRead(results);
+            logEvent(analytics, "read_sermon", {
+              sermon_date: date,
+              sermon_title: results[0]?.title,
+            });
+          }
+        } catch (err) {
+          if (isMountedRef.current) {
+            console.error("Fetch error:", err);
+            setError("Failed to load the sermon. Please try again later.");
+          }
+        } finally {
+          if (isMountedRef.current) {
+            setIsLoading(false);
+          }
+        }
+      }
     };
-  }, [location?.state?.date, location?.state?.ref]);
 
+    fetchSermon();
+
+    return () => {
+      isMountedRef.current = false;
+    };
+  }, [date, ref, navigate]);
+
+  // Effect to handle scrolling after messages are set
+  useEffect(() => {
+    if (messageToRead.length > 0 && ref) {
+      executeScroll(ref);
+    } else if (messageToRead.length > 0) {
+      // If ref is not provided, scroll to the first paragraph
+      executeScroll(1);
+    }
+  }, [messageToRead, ref]);
+
+  // Function to scroll to a specific paragraph
   const executeScroll = (i) => {
-    let element;
-    if (i) {
-      element = document.getElementById(i);
+    const element = document.getElementById(i) || document.getElementById(1);
+
+    if (element) {
+      element.scrollIntoView({
+        behavior: "smooth",
+        block: "center",
+      });
     } else {
-      element = document.getElementById(1);
+      console.warn(`Element with id "${i}" not found.`);
     }
-    element.scrollIntoView({
-      behavior: "smooth",
-      block: "center",
-    });
   };
 
-  const addSermonInfo = () => {
-    //Get the selected text and append the extra info
-    const selection = window.getSelection();
-    let sermonInfo = "";
-    let copytext = "";
-    let newdiv;
-    if (
-      messageToRead[0] &&
-      !selection.toString().includes(messageToRead[0].sermonDate)
-    ) {
-      sermonInfo =
-        messageToRead[0]?.sermonDate + " | " + messageToRead[0]?.sermonTitle;
-    }
+  // Handle copy event with proper cleanup
+  useEffect(() => {
+    const addSermonInfo = (event) => {
+      const selection = window.getSelection();
+      let sermonInfo = "";
+      let copyText = "";
 
-    copytext = sermonInfo + "\n" + selection;
+      if (messageToRead[0] && !selection.toString().includes(messageToRead[0].sermonDate)) {
+        sermonInfo = `${messageToRead[0]?.sermonDate} | ${messageToRead[0]?.sermonTitle}`;
+      }
 
-    //Create a new div to hold the prepared text
-    newdiv = document.createElement("div");
+      copyText = `${sermonInfo}\n${selection}`;
 
-    //hide the newly created container
-    newdiv.style.position = "absolute";
-    newdiv.style.left = "-99999px";
-    newdiv.style.whiteSpace = "pre-line";
+      const tempDiv = document.createElement("div");
+      tempDiv.style.position = "absolute";
+      tempDiv.style.left = "-99999px";
+      tempDiv.style.whiteSpace = "pre-line";
+      tempDiv.textContent = copyText;
 
-    //insert the container, fill it with the extended text, and define the new selection
-    document.body.appendChild(newdiv);
-    newdiv.innerHTML = copytext;
-    selection.selectAllChildren(newdiv);
+      document.body.appendChild(tempDiv);
+      selection.selectAllChildren(tempDiv);
 
-    window.setTimeout(function () {
-      document.body.removeChild(newdiv);
-    }, 100);
-  };
-  document.addEventListener("copy", addSermonInfo);
+      window.setTimeout(() => {
+        document.body.removeChild(tempDiv);
+      }, 100);
+    };
 
-  const showSermonFromSemanticSearch = () => {
-    return messageToRead?.map((entry) => (
-      <li
-        key={entry.paragraph}
-        id={entry.paragraph}
-        className={+location?.state?.ref === entry.paragraph ? "highlight" : ""}
-      >
-        <p>
-          <span className="index">{entry.paragraph}</span> {entry.section}
-        </p>
-      </li>
-    ));
-  };
+    document.addEventListener("copy", addSermonInfo);
 
-  const showSermonFromAllWordsOrExactSearch = () => {
-    return messageToRead?.map((entry) =>
-      entry.paragraph === location?.state?.ref ? (
-        <li key={entry.paragraph} id={entry.paragraph} className="highlight">
-          <p>
-            <span className="index">{entry.paragraph} </span>
-            <Highlighter
-              activeStyle={{
-                backgroundColor: "yellow",
-                color: "black",
-              }}
-              searchWords={location?.state?.searchTerm?.split(" ")}
-              autoEscape={true}
-              textToHighlight={entry.section}
-            />
-          </p>
-        </li>
-      ) : (
-        <li key={entry.paragraph} id={entry.paragraph}>
+    // Cleanup event listener on unmount
+    return () => {
+      document.removeEventListener("copy", addSermonInfo);
+    };
+  }, [messageToRead]);
+
+  // Function to render the sermon list
+  const renderSermonList = () => {
+    if (searchType === "semantic") {
+      return messageToRead.map((entry) => (
+        <li
+          key={entry.paragraph}
+          id={entry.paragraph}
+          className={+ref === entry.paragraph ? "highlight" : ""}
+        >
           <p>
             <span className="index">{entry.paragraph}</span> {entry.section}
           </p>
         </li>
-      )
-    );
+      ));
+    } else {
+      return messageToRead.map((entry) =>
+        entry.paragraph === ref ? (
+          <li key={entry.paragraph} id={entry.paragraph} className="highlight">
+            <p>
+              <span className="index">{entry.paragraph} </span>
+              <Highlighter
+                activeStyle={{
+                  backgroundColor: "yellow",
+                  color: "black",
+                }}
+                searchWords={searchTerm.split(" ")}
+                autoEscape={true}
+                textToHighlight={entry.section}
+              />
+            </p>
+          </li>
+        ) : (
+          <li key={entry.paragraph} id={entry.paragraph}>
+            <p>
+              <span className="index">{entry.paragraph}</span> {entry.section}
+            </p>
+          </li>
+        )
+      );
+    }
   };
 
   return (
-    <section className="container">
+    <section className="read-container">
       <div className="read__container">
-        {messageToRead[0]?.sermonDate || messageToRead[0]?.sermonTitle ? (
-          <h1 className="message-title">
-            {messageToRead[0]?.sermonDate} | {messageToRead[0]?.sermonTitle}
-          </h1>
-        ) : (
-          <></>
+        {isLoading && <ReadSkeleton />}
+        {error && <div className="error">{error}</div>}
+        {!isLoading && !error && (
+          <>
+            {(messageToRead[0]?.sermonDate || messageToRead[0]?.sermonTitle) && (
+              <h1 className="read-message-title">
+                {messageToRead[0]?.sermonDate} | {messageToRead[0]?.sermonTitle}
+              </h1>
+            )}
+            <ul>{renderSermonList()}</ul>
+            <ScrollToTop showUnder={260}>
+              <ScrollUpButton aria-label="scroll to the top" />
+            </ScrollToTop>
+          </>
         )}
-        <ul>
-          {location?.state?.searchType === "semantic"
-            ? showSermonFromSemanticSearch()
-            : showSermonFromAllWordsOrExactSearch()}
-        </ul>
-        <ScrollToTop showUnder={260}>
-          <ScrollUpButton aria-label="scroll to the top"></ScrollUpButton>
-        </ScrollToTop>
       </div>
     </section>
   );
